backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pillow==11.1.0
numpy==1.26.4
torch==2.6.0
tensorflow==2.16.1
python-dotenv==1.0.1
supabase==2.28.3
httpx==0.28.1
PyJWT==2.10.1
bcrypt==4.1.3
email-validator==2.1.1
websockets==15.0
pytest==8.3.5
